// src/pages/main-menu/payment/components/share-to-other/components/dialog/balanceShareHistory/models/interfaces.ts
export interface BalShareToOtherHis {
  recCreatedDate: string;
  ceilLimit?: string;
  effDate: string;
  prefix: string;
  acctResName?: string;
  isCurrency: string;
  accNbr: string;
  operationType: string;
  paymentForce: string;
  priority?: string;
  dailyCeilLimit?: string;
  expDate?: string;
  shareType?: string;
  comments?: string;
  operator?: string;
}